Module Logs_async

Async logging.

The log functions of this module return Async threads that proceed only when the log operation is over, as defined by the current Logs.reporter.

See a cooperative reporter example.

1.3 - homepage

Log functions

type 'a log = ('a, unit Async_kernel.Deferred.t) Logs.msgf -> unit Async_kernel.Deferred.t

The type for Async log functions. The returned thread only proceeds once the log operation is over. See Logs.log.

val msg : ?src:Logs.src -> Logs.level -> 'a log

See Logs.msg.

val app : ?src:Logs.src -> 'a log

See Logs.app.

val err : ?src:Logs.src -> 'a log

See Logs.err.

val warn : ?src:Logs.src -> 'a log

See Logs.warn.

val info : ?src:Logs.src -> 'a log

See Logs.info.

val debug : ?src:Logs.src -> 'a log

See Logs.debug.

val kmsg : (unit -> 'b Async_kernel.Deferred.t) -> ?src:Logs.src -> Logs.level -> ('a, 'b Async_kernel.Deferred.t) Logs.msgf -> 'b Async_kernel.Deferred.t

See Logs.kmsg.

Logging value Errors

val on_error : ?src:Logs.src -> ?level:Logs.level -> ?header:string -> ?tags:Logs.Tag.set -> pp:(Stdlib.Format.formatter -> 'b -> unit) -> use:('b -> 'a Async_kernel.Deferred.t) -> ('a, 'b) Stdlib.result Async_kernel.Deferred.t -> 'a Async_kernel.Deferred.t

See Logs.on_error.

val on_error_msg : ?src:Logs.src -> ?level:Logs.level -> ?header:string -> ?tags:Logs.Tag.set -> use:(unit -> 'a Async_kernel.Deferred.t) -> ('a, [ `Msg of string ]) Stdlib.result Async_kernel.Deferred.t -> 'a Async_kernel.Deferred.t

See Logs.on_error_msg.

Source specific log functions

module type LOG = sig ... end

val src_log : Logs.src -> (module LOG)

src_log src is a set of logging functions for src.